Add support for specifying git(hub) creds via env vars
This adds a git credential helper that uses the files stored in the
env_dir as the source. Files that match the pattern of
`GO_GIT_CRED__<PROTOCOL>__<HOSTNAME>` are considered to contain either
only the username or the username & password seperated by a colon `:`
required to authenticate to the service.

This is somewhat naive and doesn't really support anything other than
http(s) atm.

It has only been tested with Github and a personal access token so far.